+++ src/services/APIService.ts
@@ -1,6 +1,9 @@
 import { CONSTANTS } from '../utils/constants';
-import { utils } from '../utils/helpers';
 import { ragDatabase } from '../db/EnhancedVectorDatabase';
+import { fetchCVEData, processCVEData } from './nvdService';
+import { fetchEPSSData } from './epssService';
+import { fetchAIThreatIntelligence, performHeuristicAnalysis } from './aiThreatIntelService';
+import { generateAIAnalysis } from './geminiService';
 
 // Enhanced API Service Layer with Multi-Source Intelligence
 export class APIService {
@@ -23,676 +26,11 @@ export class APIService {
     }
   }
 
-  static async fetchCVEData(cveId, apiKey, setLoadingSteps) {
-    const updateSteps = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-    updateSteps(prev => [...prev, `🔍 Fetching ${cveId} from NVD...`]);
-
-    const url = `${CONSTANTS.API_ENDPOINTS.NVD}?cveId=${cveId}`;
-    const headers = {
-      'Accept': 'application/json',
-      'User-Agent': 'VulnerabilityIntelligence/1.0'
-    };
-
-    if (apiKey) headers['apiKey'] = apiKey;
-
-    const response = await this.fetchWithFallback(url, { headers });
-
-    if (!response.ok) {
-      if (response.status === 403) {
-        throw new Error('NVD API rate limit exceeded. Consider adding an API key.');
-      }
-      throw new Error(`NVD API error: ${response.status}`);
-    }
-
-    const data = await response.json();
-
-    if (!data.vulnerabilities?.length) {
-      throw new Error(`CVE ${cveId} not found in NVD database`);
-    }
-
-    updateSteps(prev => [...prev, `✅ Retrieved ${cveId} from NVD`]);
-
-    const processedData = this.processCVEData(data.vulnerabilities[0].cve);
-
-    // Store in RAG database
-    if (ragDatabase.initialized) {
-      await ragDatabase.addDocument(
-        `CVE ${cveId} NVD Data: ${processedData.description} CVSS Score: ${processedData.cvssV3?.baseScore || 'N/A'} Severity: ${processedData.cvssV3?.baseSeverity || 'Unknown'}`,
-        {
-          title: `NVD Data - ${cveId}`,
-          category: 'nvd-data',
-          tags: ['nvd', cveId.toLowerCase(), 'official-data'],
-          source: 'nvd-api',
-          cveId: cveId
-        }
-      );
-    }
-
-    return processedData;
-  }
-
-  static processCVEData(cve) {
-    const description = cve.descriptions?.find(d => d.lang === 'en')?.value || 'No description available';
-    const cvssV31 = cve.metrics?.cvssMetricV31?.[0]?.cvssData;
-    const cvssV30 = cve.metrics?.cvssMetricV30?.[0]?.cvssData;
-    const cvssV2 = cve.metrics?.cvssMetricV2?.[0]?.cvssData;
-    const cvssV3 = cvssV31 || cvssV30;
-
-    return {
-      id: cve.id,
-      description,
-      publishedDate: cve.published,
-      lastModifiedDate: cve.lastModified,
-      cvssV3: cvssV3 ? {
-        baseScore: cvssV3.baseScore,
-        baseSeverity: cvssV3.baseSeverity,
-        vectorString: cvssV3.vectorString,
-        exploitabilityScore: cvssV3.exploitabilityScore,
-        impactScore: cvssV3.impactScore,
-        attackVector: cvssV3.attackVector,
-        attackComplexity: cvssV3.attackComplexity,
-        privilegesRequired: cvssV3.privilegesRequired,
-        userInteraction: cvssV3.userInteraction,
-        scope: cvssV3.scope,
-        confidentialityImpact: cvssV3.confidentialityImpact,
-        integrityImpact: cvssV3.integrityImpact,
-        availabilityImpact: cvssV3.availabilityImpact
-      } : null,
-      cvssV2: cvssV2 ? {
-        baseScore: cvssV2.baseScore,
-        vectorString: cvssV2.vectorString,
-        accessVector: cvssV2.accessVector,
-        accessComplexity: cvssV2.accessComplexity,
-        authentication: cvssV2.authentication
-      } : null,
-      references: cve.references?.map(ref => ({
-        url: ref.url,
-        source: ref.source || 'Unknown',
-        tags: ref.tags || []
-      })) || []
-    };
-  }
-
-  static async fetchEPSSData(cveId, setLoadingSteps) {
-    const updateSteps = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-    updateSteps(prev => [...prev, `📊 Fetching EPSS data for ${cveId}...`]);
-
-    const url = `${CONSTANTS.API_ENDPOINTS.EPSS}?cve=${cveId}`;
-    const response = await this.fetchWithFallback(url, {
-      headers: {
-        'Accept': 'application/json',
-        'User-Agent': 'VulnerabilityIntelligence/1.0'
-      }
-    });
-
-    if (!response.ok) {
-      if (response.status === 404) {
-        updateSteps(prev => [...prev, `⚠️ No EPSS data available for ${cveId}`]);
-        return null;
-      }
-      throw new Error(`EPSS API error: ${response.status}`);
-    }
-
-    const data = await response.json();
-
-    if (!data.data?.length) {
-      updateSteps(prev => [...prev, `⚠️ No EPSS data found for ${cveId}`]);
-      return null;
-    }
-
-    const epssData = data.data[0];
-    const epssScore = parseFloat(epssData.epss);
-    const percentileScore = parseFloat(epssData.percentile);
-    const epssPercentage = (epssScore * 100).toFixed(3);
-
-    updateSteps(prev => [...prev, `✅ Retrieved EPSS data for ${cveId}: ${epssPercentage}% (Percentile: ${percentileScore.toFixed(3)})`]);
-
-    // Store in RAG database
-    if (ragDatabase.initialized) {
-      await ragDatabase.addDocument(
-        `CVE ${cveId} EPSS Analysis: Exploitation probability ${epssPercentage}% (percentile ${percentileScore.toFixed(3)}). ${epssScore > 0.5 ? 'High exploitation likelihood - immediate attention required.' : epssScore > 0.1 ? 'Moderate exploitation likelihood - monitor closely.' : 'Lower exploitation likelihood but monitoring recommended.'}`,
-        {
-          title: `EPSS Analysis - ${cveId}`,
-          category: 'epss-data',
-          tags: ['epss', 'exploitation-probability', cveId.toLowerCase()],
-          source: 'first-api',
-          cveId: cveId
-        }
-      );
-    }
-
-    return {
-      cve: cveId,
-      epss: epssScore.toFixed(9).substring(0, 10),
-      percentile: percentileScore.toFixed(9).substring(0, 10),
-      epssFloat: epssScore,
-      percentileFloat: percentileScore,
-      epssPercentage: epssPercentage,
-      date: epssData.date,
-      model_version: data.model_version
-    };
-  }
-
-  static async fetchAIThreatIntelligence(cveId, cveData, epssData, settings, setLoadingSteps) {
-    const updateSteps = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-
-    if (!settings.geminiApiKey) {
-      throw new Error('Gemini API key required for AI-powered threat intelligence');
-    }
-
-    const model = settings.geminiModel || 'gemini-2.5-flash';
-    const isWebSearchCapable = model.includes('2.0') || model.includes('2.5');
-
-    if (!isWebSearchCapable) {
-      updateSteps(prev => [...prev, `⚠️ Model ${model} doesn't support web search - using heuristic analysis`]);
-      return await this.performHeuristicAnalysis(cveId, cveData, epssData, setLoadingSteps);
-    }
-
-    updateSteps(prev => [...prev, `🤖 AI searching web for real-time ${cveId} threat intelligence...`]);
-
-    const searchPrompt = `You are a cybersecurity analyst researching ${cveId}. Use web search to find current information.
-
-SEARCH FOR AND ANALYZE:
-1. CISA KEV Status: Search "CISA Known Exploited Vulnerabilities ${cveId}" - Is this CVE listed in the CISA KEV catalog? If yes, get exact details including due date and required actions.
-
-2. Active Exploitation: Search "${cveId} active exploitation in the wild ransomware APT" - Are there confirmed reports of this vulnerability being actively exploited by threat actors?
-
-3. Public Exploits Discovery:
-   - Search "${cveId} exploit proof of concept github metasploit"
-   - Search "${cveId} working exploit code demonstration"
-   - Search "${cveId} exploit-db vulnerability scanner modules"
-   - Find evidence of exploits but DO NOT include URLs unless you can verify they are real GitHub repos or exploit-db entries
-
-4. Vendor Security Advisories:
-   - Search "${cveId} security advisory Microsoft Adobe Oracle"
-   - Search "${cveId} vendor patch security bulletin"
-   - Find vendor names and patch status but DO NOT generate advisory URLs
-
-5. Threat Intelligence & IOCs: Search "${cveId} threat intelligence IOCs indicators compromise" - Any IOCs, attack patterns, or threat actor usage?
-
-6. CVE Validation & False Positive Analysis:
-   - Search "${cveId} false positive disputed invalid vulnerability"
-   - Search "${cveId} vendor dispute CVE rejection withdrawn"
-   - Search "${cveId} security researcher validation confirmation"
-   - Analyze if this is a legitimate vulnerability or false positive
-
-7. Technical Analysis:
-   - Search "${cveId} technical analysis root cause impact"
-   - Search "${cveId} vulnerability details exploitation method"
-
-**CRITICAL URL HANDLING RULES**:
-1. DO NOT generate or guess any URLs
-2. DO NOT include URLs in the response unless you found them in actual search results
-3. For vendor advisories: Only provide the vendor name (e.g., "Microsoft", "Oracle", "Red Hat")
-4. For exploits: Only include GitHub URLs if you found actual repositories, otherwise just mention the source name
-5. For patches: Only include download URLs if explicitly found in search results
-6. For validation sources: Only list source names like "NVD", "Mitre", "SecurityFocus", blog names, etc.
-7. If you mention a blog or news site, just use its name without any URL
-
-CURRENT CVE DATA:
-- CVE: ${cveId}
-- CVSS: ${cveData?.cvssV3?.baseScore || 'Unknown'} (${cveData?.cvssV3?.baseSeverity || 'Unknown'})
-- EPSS: ${epssData?.epssPercentage || 'Unknown'}%
-- Attack Vector: ${cveData?.cvssV3?.attackVector || 'Unknown'}
-- Attack Complexity: ${cveData?.cvssV3?.attackComplexity || 'Unknown'}
-- Description: ${cveData?.description?.substring(0, 300) || 'No description'}
-
-For each search result, provide:
-- Source credibility (CISA, vendor, security researcher, etc.)
-- Specific findings with dates and URLs
-- Exploitation status (confirmed/suspected/none)
-- Available exploits with specific URLs and types
-- Vendor advisories and patch information with URLs
-- Validation status and any disputes
-- Recommended actions
-
-**IMPORTANT**: Actively search for and discover:
-- Specific exploit repositories and POC code
-- Vendor security advisories and patches
-- Technical analysis and validation studies
-- Any CVE disputes or false positive claims
-- Real-world exploitation evidence
-
-**CRITICAL URL HANDLING RULES**:
-- For "sources" arrays, only provide source/vendor names, NOT URLs (e.g., "Microsoft", "Red Hat", "CISA")
-- For specific advisory URLs, only include them if you can verify they are real and working
-- Do NOT generate or guess URLs - if you don't have a real URL, leave the url field empty or use ""
-- For exploit URLs, only include actual GitHub repos or exploit-db links you found
-- For vendor advisory URLs, only include if you found the actual advisory page
-
-Return your findings in this enhanced JSON structure:
-{
-  "cisaKev": {
-    "listed": boolean,
-    "details": "string with specifics including due dates",
-    "dueDate": "if applicable",
-    "source": "URL or source",
-    "emergencyDirective": boolean,
-    "aiDiscovered": true
-  },
-  "activeExploitation": {
-    "confirmed": boolean,
-    "details": "description of exploitation with evidence",
-    "sources": ["array of source URLs"],
-    "threatActors": ["known threat groups using this"],
-    "campaigns": ["specific attack campaigns"],
-    "aiDiscovered": true
-  },
-  "exploitDiscovery": {
-    "found": boolean,
-    "totalCount": number,
-    "exploits": [
-      {
-        "type": "POC/Working/Weaponized",
-        "url": "ONLY include if you found a real GitHub repo or exploit-db URL, otherwise empty string",
-        "source": "GitHub/Exploit-DB/Metasploit/etc NAME ONLY",
-        "description": "brief description",
-        "reliability": "HIGH/MEDIUM/LOW",
-        "dateFound": "discovery date"
-      }
-    ],
-    "githubRepos": number,
-    "exploitDbEntries": number,
-    "metasploitModules": number,
-    "confidence": "HIGH/MEDIUM/LOW",
-    "aiDiscovered": true
-  },
-  "vendorAdvisories": {
-    "found": boolean,
-    "count": number,
-    "advisories": [
-      {
-        "vendor": "vendor name ONLY (e.g., Microsoft, Red Hat, Oracle)",
-        "title": "advisory title if found",
-        "url": "", // LEAVE EMPTY - frontend will map to correct URL
-        "patchAvailable": boolean,
-        "patchUrl": "", // LEAVE EMPTY unless you found actual download link
-        "severity": "vendor severity rating",
-        "publishDate": "date"
-      }
-    ],
-    "patchStatus": "available/pending/none",
-    "aiDiscovered": true
-  },
-  "cveValidation": {
-    "isValid": boolean,
-    "confidence": "HIGH/MEDIUM/LOW",
-    "validationSources": ["list of source NAMES only - NVD, Mitre, Red Hat, Krebs on Security, etc. NO URLS"],
-    "disputes": [
-      {
-        "source": "who disputed",
-        "reason": "why disputed",
-        "url": "", // LEAVE EMPTY
-        "date": "dispute date"
-      }
-    ],
-    "falsePositiveIndicators": ["list of FP indicators"],
-    "legitimacyEvidence": ["evidence supporting validity"],
-    "recommendation": "VALID/FALSE_POSITIVE/DISPUTED/NEEDS_VERIFICATION",
-    "aiDiscovered": true
-  },
-  "technicalAnalysis": {
-    "rootCause": "technical root cause",
-    "exploitMethod": "how it's exploited",
-    "impactAnalysis": "detailed impact",
-    "mitigations": ["list of mitigations"],
-    "sources": ["technical analysis URLs"],
-    "aiDiscovered": true
-  },
-  "threatIntelligence": {
-    "iocs": ["any IOCs found"],
-    "threatActors": ["any associated groups"],
-    "campaignDetails": "if part of broader campaign",
-    "ransomwareUsage": boolean,
-    "aptGroups": ["nation-state actors"],
-    "aiDiscovered": true
-  },
-  "intelligenceSummary": {
-    "sourcesAnalyzed": number,
-    "exploitsFound": number,
-    "vendorAdvisoriesFound": number,
-    "activeExploitation": boolean,
-    "cisaKevListed": boolean,
-    "cveValid": boolean,
-    "threatLevel": "CRITICAL/HIGH/MEDIUM/LOW",
-    "dataFreshness": "timestamp or freshness indicator",
-    "analysisMethod": "AI_WEB_SEARCH",
-    "confidenceLevel": "HIGH/MEDIUM/LOW",
-    "aiEnhanced": true
-  },
-  "overallThreatLevel": "CRITICAL/HIGH/MEDIUM/LOW",
-  "lastUpdated": "current date",
-  "summary": "comprehensive executive summary with actionable intelligence"
-}
-
-IMPORTANT:
-- Do NOT include citation numbers like [1], [2], [3] or any bracketed numbers in your responses
-- Write all text in natural language without any citation markers
-- Focus on clear, actionable intelligence without reference numbers
-- CRITICAL: Do NOT generate any URLs. Only include URLs if you found them in actual search results and can verify they are real
-- For all sources, advisories, and validation sources, provide NAMES ONLY (no URLs)
-- The frontend will map source names to appropriate URLs to avoid 404 errors`;
-
-    try {
-      const requestBody = {
-        contents: [{
-          parts: [{ text: searchPrompt }]
-        }],
-        generationConfig: {
-          temperature: 0.1,
-          topK: 1,
-          topP: 0.95,
-          maxOutputTokens: 8192,
-          candidateCount: 1
-        },
-        tools: [{
-          google_search: {}
-        }]
-      };
-
-      const response = await this.fetchWithFallback(
-        `${CONSTANTS.API_ENDPOINTS.GEMINI}/${model}:generateContent?key=${settings.geminiApiKey}`,
-        {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(requestBody)
-        }
-      );
-
-      if (!response.ok) {
-        const errorData = await response.json().catch(() => ({}));
-        throw new Error(`AI Threat Intelligence API error: ${response.status} - ${errorData.error?.message || 'Unknown error'}`);
-      }
-
-      const data = await response.json();
-      const aiResponse = data.candidates[0].content.parts[0].text;
-
-      const updateStepsAI = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-      updateStepsAI(prev => [...prev, `✅ AI completed web-based threat intelligence analysis for ${cveId}`]);
-
-      const findings = this.parseAIThreatIntelligence(aiResponse, cveId, setLoadingSteps);
-
-      if (ragDatabase.initialized) {
-        await ragDatabase.addDocument(
-          `AI Web-Based Threat Intelligence for ${cveId}: CISA KEV: ${findings.cisaKev.listed ? 'LISTED' : 'Not Listed'}, Active Exploitation: ${findings.activeExploitation.confirmed ? 'CONFIRMED' : 'None'}, Public Exploits: ${findings.exploitDiscovery?.totalCount || findings.publicExploits?.count || 0}, Threat Level: ${findings.overallThreatLevel}. ${findings.summary}`,
-          {
-            title: `AI Web Threat Intelligence - ${cveId}`,
-            category: 'ai-web-intelligence',
-            tags: ['ai-web-search', 'threat-intelligence', cveId.toLowerCase()],
-            source: 'gemini-web-search'
-          }
-        );
-      }
-
-      return findings;
-
-    } catch (error) {
-      console.error('AI Threat Intelligence error:', error);
-      const updateStepsError = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-      updateStepsError(prev => [...prev, `⚠️ AI web search failed: ${error.message} - using fallback analysis`]);
-
-      return await this.performHeuristicAnalysis(cveId, cveData, epssData, setLoadingSteps);
-    }
-  }
-
-  static parseAIThreatIntelligence(aiResponse, cveId, setLoadingSteps) {
-    const updateStepsParse = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-
-    try {
-      const jsonMatch = aiResponse.match(/\{[\s\S]*\}/);
-      if (jsonMatch) {
-        const parsed = JSON.parse(jsonMatch[0]);
-        updateStepsParse(prev => [...prev, `📊 Parsed structured threat intelligence for ${cveId}`]);
-        return parsed;
-      }
-    } catch (e) {
-      console.log('Failed to parse JSON, analyzing text response...');
-    }
-
-    const findings = {
-      cisaKev: { listed: false, details: '', source: '', aiDiscovered: true },
-      activeExploitation: { confirmed: false, details: '', sources: [], aiDiscovered: true },
-      exploitDiscovery: {
-        found: false,
-        totalCount: 0,
-        exploits: [],
-        githubRepos: 0,
-        exploitDbEntries: 0,
-        metasploitModules: 0,
-        confidence: 'LOW',
-        aiDiscovered: true
-      },
-      publicExploits: { found: false, count: 0, sources: [], types: [] },
-      vendorAdvisories: {
-        found: false,
-        count: 0,
-        advisories: [],
-        patchStatus: 'unknown',
-        aiDiscovered: true
-      },
-      cveValidation: {
-        isValid: true,
-        confidence: 'MEDIUM',
-        validationSources: [],
-        disputes: [],
-        falsePositiveIndicators: [],
-        legitimacyEvidence: [],
-        recommendation: 'NEEDS_VERIFICATION',
-        aiDiscovered: true
-      },
-      technicalAnalysis: {
-        rootCause: '',
-        exploitMethod: '',
-        impactAnalysis: '',
-        mitigations: [],
-        sources: [],
-        aiDiscovered: true
-      },
-      threatIntelligence: {
-        iocs: [],
-        threatActors: [],
-        campaignDetails: '',
-        ransomwareUsage: false,
-        aptGroups: [],
-        aiDiscovered: true
-      },
-      intelligenceSummary: {
-        sourcesAnalyzed: 2,
-        exploitsFound: 0,
-        vendorAdvisoriesFound: 0,
-        activeExploitation: false,
-        cisaKevListed: false,
-        cveValid: true,
-        threatLevel: 'MEDIUM',
-        dataFreshness: new Date().toISOString(),
-        analysisMethod: 'HEURISTIC_FALLBACK',
-        confidenceLevel: 'LOW',
-        aiEnhanced: false
-      },
-      overallThreatLevel: 'MEDIUM',
-      lastUpdated: new Date().toISOString(),
-      summary: 'AI analysis completed with limited results'
-    };
-
-    const response = aiResponse.toLowerCase();
-
-    if (response.includes('cisa kev') || response.includes('known exploited')) {
-      if (response.includes('listed') || response.includes('catalog')) {
-        findings.cisaKev.listed = true;
-        findings.cisaKev.details = 'Found in CISA Known Exploited Vulnerabilities catalog';
-        findings.overallThreatLevel = 'CRITICAL';
-      }
-    }
-
-    if (response.includes('active exploit') || response.includes('in the wild')) {
-      findings.activeExploitation.confirmed = true;
-      findings.activeExploitation.details = 'Active exploitation detected in threat intelligence';
-      findings.overallThreatLevel = 'HIGH';
-    }
-
-    if (response.includes('exploit') && (response.includes('github') || response.includes('poc'))) {
-      findings.exploitDiscovery.found = true;
-      findings.exploitDiscovery.totalCount = (response.match(/exploit/g) || []).length;
-      findings.exploitDiscovery.confidence = 'MEDIUM';
-      findings.publicExploits.found = true;
-      findings.publicExploits.count = findings.exploitDiscovery.totalCount;
-      findings.publicExploits.types = ['POC'];
-    }
-
-    const urls = aiResponse.match(/https?:\/\/[^\s<>"{}|\\^`[\]]+/g);
-    if (urls) {
-      findings.activeExploitation.sources = urls.slice(0, 3);
-      findings.exploitDiscovery.exploits = urls.slice(0, 5).map((url, idx) => ({
-        type: 'POC',
-        url: url,
-        source: 'Web Search',
-        description: 'Found via AI web search',
-        reliability: 'MEDIUM',
-        dateFound: new Date().toISOString()
-      }));
-      findings.publicExploits.sources = urls.slice(0, 5);
-    }
-
-    findings.intelligenceSummary.exploitsFound = findings.exploitDiscovery.totalCount;
-    findings.intelligenceSummary.activeExploitation = findings.activeExploitation.confirmed;
-    findings.intelligenceSummary.cisaKevListed = findings.cisaKev.listed;
-
-    findings.summary = `AI web search analysis: ${findings.cisaKev.listed ? 'CISA KEV listed' : 'Not in KEV'}, ${findings.activeExploitation.confirmed ? 'Active exploitation' : 'No active exploitation'}, ${findings.exploitDiscovery.totalCount} potential exploits found`;
-
-    const updateStepsSummary = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-    updateStepsSummary(prev => [...prev, `📈 AI analysis: ${findings.overallThreatLevel} threat level determined`]);
-
-    return findings;
-  }
-
-  static async performHeuristicAnalysis(cveId, cveData, epssData, setLoadingSteps) {
-    const updateStepsHeuristic = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-    updateStepsHeuristic(prev => [...prev, `🔍 Performing advanced heuristic analysis for ${cveId}...`]);
-
-    const year = parseInt(cveId.split('-')[1]);
-    const id = parseInt(cveId.split('-')[2]);
-    const cvssScore = cveData?.cvssV3?.baseScore || cveData?.cvssV2?.baseScore || 0;
-    const epssFloat = epssData?.epssFloat || 0;
-    const severity = utils.getSeverityLevel(cvssScore);
-
-    let riskScore = 0;
-    const indicators = [];
-
-    if (cvssScore >= 9) { riskScore += 4; indicators.push('Critical CVSS score'); }
-    else if (cvssScore >= 7) { riskScore += 3; indicators.push('High CVSS score'); }
-
-    if (epssFloat > 0.7) { riskScore += 4; indicators.push('Very high EPSS score'); }
-    else if (epssFloat > 0.3) { riskScore += 2; indicators.push('Elevated EPSS score'); }
-
-    if (year >= 2024) { riskScore += 2; indicators.push('Recent vulnerability'); }
-    if (id < 1000) { riskScore += 2; indicators.push('Early discovery in year'); }
-
-    const highRiskPatterns = ['21413', '44487', '38030', '26923', '1675'];
-    if (highRiskPatterns.some(pattern => cveId.includes(pattern))) {
-      riskScore += 5;
-      indicators.push('Matches known high-risk pattern');
-    }
-
-    const description = cveData?.description?.toLowerCase() || '';
-    const highValueTargets = ['microsoft', 'apache', 'oracle', 'vmware', 'cisco', 'windows', 'exchange', 'linux'];
-    if (highValueTargets.some(target => description.includes(target))) {
-      riskScore += 2;
-      indicators.push('Affects high-value target software');
-    }
-
-    const threatLevel = riskScore >= 8 ? 'CRITICAL' : riskScore >= 6 ? 'HIGH' : riskScore >= 4 ? 'MEDIUM' : 'LOW';
-    const likelyInKEV = riskScore >= 7;
-    const likelyExploited = riskScore >= 5;
-    const exploitCount = Math.min(Math.floor(riskScore / 2), 5);
-
-    const updateStepsComplete = typeof setLoadingSteps === 'function' ? setLoadingSteps : () => {};
-    updateStepsComplete(prev => [...prev, `📊 Heuristic analysis complete: ${threatLevel} threat level (score: ${riskScore})`]);
-
-    return {
-      cisaKev: {
-        listed: likelyInKEV,
-        details: likelyInKEV ? 'High probability of KEV listing based on risk factors' : 'Low probability of KEV listing',
-        confidence: 'HEURISTIC',
-        source: 'Advanced pattern analysis',
-        aiDiscovered: false
-      },
-      activeExploitation: {
-        confirmed: likelyExploited,
-        details: likelyExploited ? 'High exploitation likelihood based on multiple risk factors' : 'Lower exploitation probability',
-        sources: [`Risk indicators: ${indicators.join(', ')}`],
-        aiDiscovered: false
-      },
-      exploitDiscovery: {
-        found: exploitCount > 0,
-        totalCount: exploitCount,
-        exploits: exploitCount > 0 ? [{
-          type: exploitCount > 2 ? 'Working Exploit' : 'POC',
-          url: `https://www.exploit-db.com/search?cve=${cveId}`,
-          source: 'Exploit-DB (Predicted)',
-          description: 'Heuristic prediction based on vulnerability characteristics',
-          reliability: riskScore >= 6 ? 'HIGH' : 'MEDIUM',
-          dateFound: new Date().toISOString()
-        }] : [],
-        githubRepos: Math.max(0, exploitCount - 1),
-        exploitDbEntries: exploitCount > 0 ? 1 : 0,
-        metasploitModules: exploitCount > 3 ? 1 : 0,
-        confidence: riskScore >= 6 ? 'HIGH' : 'MEDIUM',
-        aiDiscovered: false
-      },
-      vendorAdvisories: {
-        found: Math.floor(riskScore / 3) > 0,
-        count: Math.floor(riskScore / 3),
-        advisories: [],
-        patchStatus: cvssScore >= 7 ? 'likely available' : 'pending',
-        aiDiscovered: false
-      },
-      cveValidation: {
-        isValid: true,
-        confidence: 'MEDIUM',
-        validationSources: ['NVD', 'EPSS'],
-        disputes: [],
-        falsePositiveIndicators: [],
-        legitimacyEvidence: indicators,
-        recommendation: 'VALID',
-        aiDiscovered: false
-      },
-      technicalAnalysis: {
-        rootCause: 'Analysis based on CVE description and scoring',
-        exploitMethod: cvssScore >= 7 ? 'Remote exploitation likely' : 'Local access may be required',
-        impactAnalysis: `${severity} impact vulnerability with ${cvssScore} CVSS score`,
-        mitigations: ['Apply vendor patches', 'Monitor for exploitation attempts', 'Implement network controls'],
-        sources: [],
-        aiDiscovered: false
-      },
-      threatIntelligence: {
-        iocs: [],
-        threatActors: [],
-        campaignDetails: riskScore >= 8 ? 'Possible APT interest due to high impact' : '',
-        ransomwareUsage: riskScore >= 7,
-        aptGroups: [],
-        aiDiscovered: false
-      },
-      intelligenceSummary: {
-        sourcesAnalyzed: 2,
-        exploitsFound: exploitCount,
-        vendorAdvisoriesFound: Math.floor(riskScore / 3),
-        activeExploitation: likelyExploited,
-        cisaKevListed: likelyInKEV,
-        cveValid: true,
-        threatLevel: threatLevel,
-        dataFreshness: new Date().toISOString(),
-        analysisMethod: 'ADVANCED_HEURISTICS',
-        confidenceLevel: riskScore >= 6 ? 'HIGH' : 'MEDIUM',
-        aiEnhanced: false
-      },
-      overallThreatLevel: threatLevel,
-      lastUpdated: new Date().toISOString(),
-      summary: `Heuristic analysis: ${indicators.length} risk indicators detected, ${threatLevel} threat level assigned`,
-      analysisMethod: 'ADVANCED_HEURISTICS',
-      riskScore: riskScore,
-      indicators: indicators
-    };
-  }
+  // Keep processCVEData here if it's used by other methods within APIService, or move it if only used by fetchCVEData
+  // For now, assuming it might be used elsewhere or is closely tied.
+  // If it's only used by fetchCVEData, it should ideally be in nvdService.ts and not exported from there if not needed externally.
+  // However, to minimize initial disruption, keeping it here if no immediate issues arise.
+  // static processCVEData = processCVEData; // Already in nvdService, no need to re-declare here if imported
 
   static async fetchVulnerabilityDataWithAI(cveId, setLoadingSteps, apiKeys, settings) {
     try {
@@ -700,14 +38,14 @@ IMPORTANT:
 
       if (!ragDatabase.initialized) {
         setLoadingSteps(prev => [...prev, `📚 Initializing RAG knowledge base...`]);
-        await ragDatabase.initialize();
+        await ragDatabase.initialize(settings.geminiApiKey); // Pass API key for initialization
       }
 
       setLoadingSteps(prev => [...prev, `🔍 Fetching from primary sources (NVD, EPSS)...`]);
 
       const [cveResult, epssResult] = await Promise.allSettled([
-        this.fetchCVEData(cveId, apiKeys.nvd, setLoadingSteps),
-        this.fetchEPSSData(cveId, setLoadingSteps)
+        fetchCVEData(cveId, apiKeys.nvd, setLoadingSteps),
+        fetchEPSSData(cveId, setLoadingSteps)
       ]);
 
       const cve = cveResult.status === 'fulfilled' ? cveResult.value : null;
@@ -719,7 +57,7 @@ IMPORTANT:
 
       setLoadingSteps(prev => [...prev, `🌐 AI analyzing real-time threat intelligence via web search...`]);
 
-      const aiThreatIntel = await this.fetchAIThreatIntelligence(cveId, cve, epss, settings, setLoadingSteps);
+      const aiThreatIntel = await fetchAIThreatIntelligence(cveId, cve, epss, settings, setLoadingSteps);
 
       const discoveredSources = ['NVD'];
       const sources = [{ name: 'NVD', url: `https://nvd.nist.gov/vuln/detail/${cveId}`, aiDiscovered: false }];
@@ -867,274 +205,11 @@ IMPORTANT:
     }
   }
 
+  // This method is now a wrapper calling the one from geminiService.ts
   static async generateAIAnalysis(vulnerability, apiKey, model, settings = {}) {
-    if (!apiKey) throw new Error('Gemini API key required');
-
-    const now = Date.now();
-    const lastRequest = window.lastGeminiRequest || 0;
-
-    if ((now - lastRequest) < CONSTANTS.RATE_LIMITS.GEMINI_COOLDOWN) {
-      const waitTime = Math.ceil((CONSTANTS.RATE_LIMITS.GEMINI_COOLDOWN - (now - lastRequest)) / 1000);
-      throw new Error(`Rate limit protection: Please wait ${waitTime} more seconds. Free Gemini API has strict limits.`);
-    }
-
-    window.lastGeminiRequest = now;
-
-    await ragDatabase.ensureInitialized(apiKey);
-    console.log(`📊 RAG Database Status: ${ragDatabase.documents.length} documents available (${ragDatabase.geminiApiKey ? 'Gemini embeddings' : 'local embeddings'})`);
-
-    const cveId = vulnerability.cve.id;
-    const ragQuery = `${cveId} ${vulnerability.cve.description.substring(0, 200)} vulnerability analysis security impact mitigation threat intelligence EPSS ${vulnerability.epss?.epssPercentage || 'N/A'} CVSS ${vulnerability.cve.cvssV3?.baseScore || 'N/A'} ${vulnerability.kev?.listed ? 'CISA KEV active exploitation' : ''}`;
-
-    console.log(`🔍 RAG Query: "${ragQuery.substring(0, 100)}..."`);
-    const relevantDocs = await ragDatabase.search(ragQuery, 15);
-    console.log(`📚 RAG Retrieved: ${relevantDocs.length} relevant documents (${relevantDocs.filter(d => d.embeddingType === 'gemini').length} with Gemini embeddings)`);
-
-    const ragContext = relevantDocs.length > 0 ?
-      relevantDocs.map((doc, index) =>
-        `[Security Knowledge ${index + 1}] ${doc.metadata.title} (Relevance: ${(doc.similarity * 100).toFixed(1)}%, ${doc.embeddingType || 'local'} embedding):\n${doc.content.substring(0, 800)}...`
-      ).join('\n\n') :
-      'No specific security knowledge found in database. Initializing knowledge base for future queries.';
-
-    if (relevantDocs.length === 0) {
-      console.log('🔄 No specific matches found, trying broader search...');
-      const broaderQuery = `vulnerability security analysis ${vulnerability.cve.cvssV3?.baseSeverity || 'unknown'} severity`;
-      const broaderDocs = await ragDatabase.search(broaderQuery, 8);
-      console.log(`📚 Broader RAG Search: ${broaderDocs.length} documents found`);
-
-      if (broaderDocs.length > 0) {
-        const broaderContext = broaderDocs.map((doc, index) =>
-          `[General Security Knowledge ${index + 1}] ${doc.metadata.title} (${doc.embeddingType || 'local'} embedding):\n${doc.content.substring(0, 600)}...`
-        ).join('\n\n');
-
-        relevantDocs.push(...broaderDocs);
-      }
-    }
-
-    const prompt = this.buildEnhancedAnalysisPrompt(vulnerability, ragContext, relevantDocs.length);
-
-    const requestBody = {
-      contents: [{ parts: [{ text: prompt }] }],
-      generationConfig: {
-        temperature: 0.1,
-        topK: 1,
-        topP: 0.8,
-        maxOutputTokens: 8192,
-        candidateCount: 1
-      }
-    };
-
-    const isWebSearchCapable = model.includes('2.0') || model.includes('2.5');
-    if (isWebSearchCapable) {
-      requestBody.tools = [{ google_search: {} }];
-    }
-
-    const apiUrl = `${CONSTANTS.API_ENDPOINTS.GEMINI}/${model}:generateContent?key=${apiKey}`;
-
-    try {
-      const response = await this.fetchWithFallback(apiUrl, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(requestBody)
-      });
-
-      if (!response.ok) {
-        const errorData = await response.json().catch(() => ({}));
-
-        if (response.status === 429) {
-          throw new Error('Gemini API rate limit exceeded. Please wait a few minutes before trying again.');
-        }
-
-        if (response.status === 401 || response.status === 403) {
-          throw new Error('Invalid Gemini API key. Please check your API key in settings.');
-        }
-
-        throw new Error(`Gemini API error: ${response.status} - ${errorData.error?.message || 'Unknown error'}`);
-      }
-
-      const data = await response.json();
-      const content = data.candidates?.[0]?.content;
-
-      if (!content?.parts?.[0]?.text) {
-        throw new Error('Invalid response from Gemini API');
-      }
-
-      const analysisText = content.parts[0].text;
-
-      if (!analysisText || analysisText.trim().length === 0) {
-        throw new Error('Empty analysis received from Gemini API');
-      }
-
-      if (analysisText.length > 500) {
-        await ragDatabase.addDocument(
-          `Enhanced CVE Analysis: ${cveId}\n\nCVSS: ${vulnerability.cve.cvssV3?.baseScore || 'N/A'}\nEPSS: ${vulnerability.epss?.epssPercentage || 'N/A'}%\nCISA KEV: ${vulnerability.kev?.listed ? 'Yes' : 'No'}\n\n${analysisText}`,
-          {
-            title: `Enhanced RAG Security Analysis - ${cveId}`,
-            category: 'enhanced-analysis',
-            tags: ['rag-enhanced', 'ai-analysis', cveId.toLowerCase(), vulnerability.cve.cvssV3?.baseSeverity?.toLowerCase() || 'unknown'],
-            source: 'ai-analysis-rag',
-            model: model,
-            cveId: cveId
-          }
-        );
-        console.log(`💾 Stored analysis for ${cveId} in RAG database for future reference`);
-      }
-
-      return {
-        analysis: analysisText,
-        ragUsed: true,
-        ragDocuments: relevantDocs.length,
-        ragSources: relevantDocs.map(doc => doc.metadata?.title || 'Unknown').filter(Boolean),
-        webGrounded: isWebSearchCapable,
-        enhancedSources: vulnerability.enhancedSources || [],
-        discoveredSources: vulnerability.discoveredSources || [],
-        model: model,
-        analysisTimestamp: new Date().toISOString(),
-        ragDatabaseSize: ragDatabase.documents.length,
-        embeddingType: ragDatabase.geminiApiKey ? 'gemini' : 'local',
-        geminiEmbeddingsCount: ragDatabase.documents.filter(d => d.embeddingType === 'gemini').length,
-        realTimeData: {
-          cisaKev: vulnerability.kev?.listed || false,
-          exploitsFound: vulnerability.exploits?.count || 0,
-          exploitConfidence: vulnerability.exploits?.confidence || 'NONE',
-          githubRefs: vulnerability.github?.count || 0,
-          threatLevel: vulnerability.threatLevel || 'STANDARD',
-          heuristicRisk: vulnerability.kev?.heuristicHighRisk || false
-        }
-      };
-
-    } catch (error) {
-      console.error('Enhanced RAG Analysis Error:', error);
-      return this.generateEnhancedFallbackAnalysis(vulnerability, error);
-    }
-  }
-
-  static buildEnhancedAnalysisPrompt(vulnerability, ragContext, ragDocCount = 0) {
-    const cveId = vulnerability.cve.id;
-    const cvssScore = vulnerability.cve.cvssV3?.baseScore || vulnerability.cve.cvssV2?.baseScore || 'N/A';
-    const epssScore = vulnerability.epss ? vulnerability.epss.epssPercentage + '%' : 'N/A';
-    const kevStatus = vulnerability.kev?.listed ? 'Yes - ACTIVE EXPLOITATION CONFIRMED' : 'No';
-
-    return `You are a senior cybersecurity analyst providing comprehensive vulnerability assessment for ${cveId}.
-
-VULNERABILITY DETAILS:
-- CVE ID: ${cveId}
-- CVSS Score: ${cvssScore}
-- EPSS Score: ${epssScore}
-- CISA KEV Status: ${kevStatus}
-- Description: ${vulnerability.cve.description.substring(0, 800)}
-
-REAL-TIME THREAT INTELLIGENCE:
-${vulnerability.kev?.listed ? `⚠️ CRITICAL: This vulnerability is actively exploited according to CISA KEV catalog.` : ''}
-${vulnerability.exploits?.found ? `💣 PUBLIC EXPLOITS: ${vulnerability.exploits.count} exploit(s) found with ${vulnerability.exploits.confidence || 'MEDIUM'} confidence.` : ''}
-${vulnerability.github?.found ? `🔍 GITHUB REFS: ${vulnerability.github.count} security-related repositories found.` : ''}
-${vulnerability.activeExploitation?.confirmed ? `🚨 ACTIVE EXPLOITATION: Confirmed exploitation in the wild.` : ''}
-
-SECURITY KNOWLEDGE BASE (${ragDocCount} relevant documents retrieved):
-${ragContext}
-
-DATA SOURCES ANALYZED:
-${vulnerability.discoveredSources?.join(', ') || 'NVD, EPSS'}
-
-You have access to ${ragDocCount} relevant security documents from the knowledge base. Use this contextual information to provide enhanced insights beyond standard vulnerability analysis.
-
-Provide a comprehensive vulnerability analysis including:
-1. Executive Summary with immediate actions needed
-2. Technical details and attack vectors
-3. Impact assessment and potential consequences
-4. Mitigation strategies and remediation guidance
-5. Affected systems and software components
-6. Current exploitation status and threat landscape
-7. Priority recommendations based on real-time threat intelligence
-8. Lessons learned from similar vulnerabilities (use knowledge base context)
-
-Format your response in clear sections with detailed analysis. Leverage the security knowledge base context and real-time threat intelligence to provide enhanced insights that go beyond basic CVE information.
-
-${vulnerability.kev?.listed ? 'EMPHASIZE THE CRITICAL NATURE DUE TO CONFIRMED ACTIVE EXPLOITATION.' : ''}
-${vulnerability.exploits?.found && vulnerability.exploits.confidence === 'HIGH' ? 'HIGHLIGHT THE AVAILABILITY OF PUBLIC EXPLOITS.' : ''}
-
-**Important**:
-- Reference insights from the security knowledge base when relevant to demonstrate enhanced RAG-powered analysis.
-- DO NOT include citation numbers like [1], [2], [3] or any bracketed numbers in your response.
-- Write in clear, natural language without any citation markers.`;
-  }
-
-  static generateEnhancedFallbackAnalysis(vulnerability, error) {
-    const cveId = vulnerability.cve.id;
-    const cvssScore = vulnerability.cve.cvssV3?.baseScore || vulnerability.cve.cvssV2?.baseScore || 'N/A';
-    const epssScore = vulnerability.epss ? vulnerability.epss.epssPercentage + '%' : 'N/A';
-    const kevStatus = vulnerability.kev?.listed ? 'Yes - ACTIVE EXPLOITATION CONFIRMED' : 'No';
-
-    return {
-      analysis: `# Security Analysis: ${cveId}
-
-## Executive Summary
-${kevStatus.includes('Yes') ? '🚨 **CRITICAL PRIORITY** - This vulnerability is actively exploited according to CISA KEV catalog. Immediate patching required.' :
-  vulnerability.exploits?.found ? `💣 **HIGH RISK** - ${vulnerability.exploits.count} exploit(s) detected with ${vulnerability.exploits.confidence} confidence level.` :
-  `This vulnerability has a CVSS score of ${cvssScore} with an EPSS exploitation probability of ${epssScore}.`}
-
-${vulnerability.exploits?.found ? `💣 **PUBLIC EXPLOITS AVAILABLE** - ${vulnerability.exploits.count} exploit(s) detected with ${vulnerability.exploits.confidence} confidence level.` : ''}
-
-## Vulnerability Details
-**CVE ID:** ${cveId}
-**CVSS Score:** ${cvssScore}
-**EPSS Score:** ${epssScore}
-**CISA KEV Status:** ${kevStatus}
-
-**Description:** ${vulnerability.cve.description}
-
-## Real-Time Threat Intelligence Summary
-${vulnerability.kev?.listed ? '- ⚠️ **ACTIVE EXPLOITATION**: Confirmed in CISA Known Exploited Vulnerabilities catalog' : '- No confirmed active exploitation in CISA KEV catalog'}
-${vulnerability.exploits?.found ? `- 💣 **PUBLIC EXPLOITS**: ${vulnerability.exploits.count} exploit(s) with ${vulnerability.exploits.confidence} confidence` : '- No high-confidence public exploits identified'}
-${vulnerability.github?.found ? `- 🔍 **SECURITY COVERAGE**: ${vulnerability.github.count} GitHub security references found` : '- Limited GitHub security advisory coverage'}
-${vulnerability.activeExploitation?.confirmed ? '- 🚨 **ACTIVE EXPLOITATION**: Confirmed exploitation detected in threat intelligence' : '- No confirmed active exploitation detected'}
-
-## Risk Assessment
-**Exploitation Probability:** ${epssScore} (EPSS)
-**Attack Vector:** ${vulnerability.cve.cvssV3?.attackVector || 'Unknown'}
-**Attack Complexity:** ${vulnerability.cve.cvssV3?.attackComplexity || 'Unknown'}
-**Privileges Required:** ${vulnerability.cve.cvssV3?.privilegesRequired || 'Unknown'}
-**Impact Level:** ${vulnerability.cve.cvssV3?.baseSeverity || 'Unknown'}
-
-## Immediate Actions Required
-1. ${kevStatus.includes('Yes') || vulnerability.exploits?.found ?
-   'URGENT: Apply patches immediately - high exploitation risk confirmed' :
-   'Review and prioritize patching based on CVSS score and environment exposure'}
-2. ${vulnerability.exploits?.found ? 'Implement additional monitoring - public exploits available' : 'Monitor for unusual activity patterns'}
-3. Review access controls and authentication mechanisms
-4. ${vulnerability.kev?.listed ? 'Follow CISA emergency directive timelines' : 'Consider temporary compensating controls if patches unavailable'}
-
-## Mitigation Strategies
-- **Patch Management**: ${kevStatus.includes('Yes') ? 'Emergency patching within CISA timeline' : 'Standard patch testing and deployment'}
-- **Network Controls**: Implement input validation and filtering
-- **Access Controls**: Review and restrict privileged access
-- **Monitoring**: Deploy detection rules for exploitation attempts
-
-## Data Sources Analyzed
-${vulnerability.discoveredSources?.join(', ') || 'NVD, EPSS'} (${vulnerability.discoveredSources?.length || 2} sources)
-
-## Intelligence Assessment
-- **Data Freshness**: Real-time (${new Date().toLocaleString()})
-- **Confidence Level**: ${vulnerability.exploits?.confidence || 'MEDIUM'} based on multiple source correlation
-- **Threat Landscape**: ${vulnerability.threatLevel || 'STANDARD'} risk environment
-
-*Analysis generated using real-time threat intelligence. Enhanced AI service temporarily unavailable due to: ${error.message}*`,
-      ragUsed: false,
-      ragDocuments: 0,
-      ragSources: [],
-      webGrounded: false,
-      enhancedSources: vulnerability.enhancedSources || [],
-      discoveredSources: vulnerability.discoveredSources || [],
-      error: error.message,
-      fallbackUsed: true,
-      realTimeData: {
-        cisaKev: vulnerability.kev?.listed || false,
-        exploitsFound: vulnerability.exploits?.count || 0,
-        exploitConfidence: vulnerability.exploits?.confidence || 'NONE',
-        githubRefs: vulnerability.github?.count || 0,
-        threatLevel: vulnerability.threatLevel || 'STANDARD',
-        activeExploitation: vulnerability.activeExploitation?.confirmed || false
-      }
-    };
+    // The actual implementation is in geminiService.generateAIAnalysis
+    // This ensures that APIService.generateAIAnalysis can still be called from components
+    // without them needing to know about geminiService directly for this specific function.
+    return generateAIAnalysis(vulnerability, apiKey, model, settings);
   }
 }
